+++ src/libs/Notification.ts
@@ -1,3 +1,5 @@
+import * as _ from 'lodash'
+import * as pluralize from 'pluralize'
 import * as shortid from 'shortid'
 
 import LogType from 'Constants/logType'
@@ -7,6 +9,104 @@ import LogType from 'Constants/logType'
  * A notification can be a cheer, a raid, a ritual, a resub, a sub, a subgift, etc.
  */
 export default class Notification implements Serializable<SerializedNotification> {
+  /**
+   * Creates a new notification from a submysterygift user notice which corresponds to a known user gifting to a random
+   * user.
+   * @param  tags - The notice tags.
+   * @return The new notice.
+   */
+  public static fromSubMysteryGift(tags: Record<string, string>) {
+    const username = tags['display-name'] || tags.login
+    const total = parseInt(tags['msg-param-mass-gift-count'], 10)
+
+    return new Notification(
+      `${username} is giving out ${total} sub ${pluralize('gift', total)}!`,
+      NotificationEvent.SubMysteryGift
+    )
+  }
+
+  /**
+   * Creates a new notification from an anonsubgift user notice which corresponds to an unknown user gifting to a
+   * specific user.
+   * @param  tags - The notice tags.
+   * @return The new notice.
+   */
+  public static fromAnonSubGift(tags: Record<string, string>) {
+    const recipient = tags['msg-param-recipient-display-name'] || tags['msg-param-recipient-user-name']
+
+    return new Notification(`An anonymous gifter just gifted a sub to ${recipient}!`, NotificationEvent.AnonSubGift)
+  }
+
+  /**
+   * Creates a new notification from an anonsubmysterygift user notice which corresponds to an unknown user gifting to
+   * multiple random users.
+   * @param  tags - The notice tags.
+   * @return The new notice.
+   */
+  public static fromAnonSubMysteryGift(tags: Record<string, string>) {
+    const total = parseInt(tags['msg-param-mass-gift-count'], 10)
+
+    return new Notification(
+      `An anonymous gifter is giving out ${total} sub ${pluralize('gift', total)}!`,
+      NotificationEvent.AnonSubMysteryGift
+    )
+  }
+
+  /**
+   * Creates a new notification from a giftpaidupgrade user notice which corresponds to a user continuing the gift sub
+   * they got from a known user.
+   * @param  tags - The notice tags.
+   * @return The new notice.
+   */
+  public static fromGiftPaidUpgrade(tags: Record<string, string>) {
+    const username = tags['display-name'] || tags.login
+    const gifter = tags['msg-param-sender-name'] || tags['msg-param-sender-login']
+
+    return new Notification(
+      `${username} is continuing the gift sub they got from ${gifter}!`,
+      NotificationEvent.GiftPaidUpgrade
+    )
+  }
+
+  /**
+   * Creates a new notification from an anongiftpaidupgrade user notice which corresponds to a user continuing the gift
+   * sub they got from an unknown user.
+   * @param  tags - The notice tags.
+   * @return The new notice.
+   */
+  public static fromAnonGiftPaidUpgrade(tags: Record<string, string>) {
+    const username = tags['display-name'] || tags.login
+
+    return new Notification(
+      `${username} is continuing the gift sub they got from an anonymous gifter!`,
+      NotificationEvent.AnonGiftPaidUpgrade
+    )
+  }
+
+  /**
+   * Creates a new notification from a primepaidupgrade user notice which corresponds to a user upgrading their Prime
+   * sub to a normal paid one.
+   * @param  tags - The notice tags.
+   * @return The new notice.
+   */
+  public static fromPrimePaidUpgrade(tags: Record<string, string>) {
+    const username = tags['display-name'] || tags.login
+    const plan = _.get(tags, 'msg-param-sub-plan', '')
+
+    let tier = 1
+
+    if (plan === '2000') {
+      tier = 2
+    } else if (plan === '3000') {
+      tier = 3
+    }
+
+    return new Notification(
+      `${username} just converted their Twitch Prime sub to a tier ${tier} sub!`,
+      NotificationEvent.PrimePaidUpgrade
+    )
+  }
+
   private id: string
 
   /**
@@ -39,12 +139,18 @@ export default class Notification implements Serializable<SerializedNotification
  * Notification associated events.
  */
 export enum NotificationEvent {
+  AnonGiftPaidUpgrade,
+  AnonSubGift,
+  AnonSubMysteryGift,
+  GiftPaidUpgrade,
   Host,
+  PrimePaidUpgrade,
   Raid,
   ReSub,
   RitualNewChatter,
   SubGift,
   Subscription,
+  SubMysteryGift,
 }
 
 /**